+++ Projects/App/Sources/Application/AppComponent+Faq.swift
@@ -7,11 +7,11 @@
 //
 
 import CommonFeature
-import DataModule
-import DomainModule
 import NetworkModule
 import SignInFeature
 import StorageFeature
+import FaqDomain
+import FaqDomainInterface
 
 // MARK: 변수명 주의
 // AppComponent 내 변수 == Dependency 내 변수  이름 같아야함
