+++ Projects/Domains/FaqDomain/Interface/DataSource/RemoteFaqDataSource.swift
@@ -1,6 +1,3 @@
-import DataMappingModule
-import DomainModule
-import ErrorModule
 import Foundation
 import RxSwift
