+++ Projects/Domains/FaqDomain/Interface/Entity/FaqCategoryEntity.swift
