+++ Projects/Domains/FaqDomain/Interface/Repository/FaqRepository.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
-import DataMappingModule
-import ErrorModule
 import Foundation
 import RxSwift
