+++ Projects/Domains/FaqDomain/Interface/UseCase/FetchFaqCategoriesUseCase.swift
@@ -6,10 +6,9 @@
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
-import DataMappingModule
 import Foundation
 import RxSwift
 
-public protocol FetchFaqUseCase {
-    func execute() -> Single<[FaqEntity]>
+public protocol FetchFaqCategoriesUseCase {
+    func execute() -> Single<FaqCategoryEntity>
 }
