+++ Projects/Domains/FaqDomain/Sources/API/FaqAPI.swift
@@ -1,8 +1,8 @@
-import DataMappingModule
 import ErrorModule
 import Foundation
 import KeychainModule
 import Moya
+import BaseDomain
 
 public enum FaqAPI {
     case fetchFaqCategories
