+++ Projects/Domains/FaqDomain/Sources/ResponseDTO/FaqCategoryResponseDTO.swift
@@ -1,6 +1,9 @@
-import DataMappingModule
-import DomainModule
-import Utility
+import Foundation
+import FaqDomainInterface
+
+public struct FaqCategoryResponseDTO: Decodable, Equatable {
+    public let categories: [String]
+}
 
 public extension FaqCategoryResponseDTO {
     func toDomain() -> FaqCategoryEntity {
