--- Projects/Domains/FaqDomain/Sources/Sources.swift
@@ -1,1 +1,0 @@
-// This is for Tuist
