+++ Projects/Features/StorageFeature/Sources/Components/FaqContentComponent.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
-import DomainModule
+import FaqDomainInterface
 import Foundation
 import NeedleFoundation
