+++ Projects/Features/StorageFeature/Sources/ViewModels/QnaContentViewModel.swift
@@ -7,12 +7,12 @@
 //
 
 import BaseFeature
-import DomainModule
 import Foundation
 import KeychainModule
 import RxRelay
 import RxSwift
 import Utility
+import FaqDomainInterface
 
 public final class QnaContentViewModel: ViewModelType {
     var disposeBag = DisposeBag()
