+++ Projects/Features/StorageFeature/Sources/Views/AnswerTableViewCell.swift
@@ -7,7 +7,7 @@
 //
 
 import DesignSystem
-import DomainModule
+import FaqDomainInterface
 import UIKit
 
 class AnswerTableViewCell: UITableViewCell {
