--- Projects/Services/DataMappingModule/Sources/Faq/Response/FaqCategoryResponseDTO.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct FaqCategoryResponseDTO: Codable, Equatable {
-    public let categories: [String]
-}
