--- Projects/Services/NetworkModule/Sources/Faq/DataTransfer/FetchQnaTransfer.swift
@@ -1,14 +1,0 @@
-import DataMappingModule
-import DomainModule
-import Utility
-
-public extension FaqResponseDTO {
-    func toDomain() -> FaqEntity {
-        FaqEntity(
-            category: category,
-            question: question,
-            description: description,
-            isOpen: false
-        )
-    }
-}
